backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
asyncpg>=0.30.0
python-dotenv>=1.0.1
orjson>=3.10.0
mlflow[mcp]==3.8.1
numpy==2.4.1
scikit-learn==1.8.0
pyarrow==22.0.0
